Added External and Internal BST along with headers
I haven't tested yet, but they should be working as intended. One thing to note is that they have the same code except insert in external is different by a few lines of code. Really shows how easy it would be to make one general class for both to use.